requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
